+++ packages/snap/src/clients/price-api/types.ts
@@ -12,6 +12,7 @@ import {
   record,
   string,
   tuple,
+  type as typeStruct,
   union,
 } from '@metamask/superstruct';
 import { CaipAssetTypeStruct } from '@metamask/utils';
@@ -150,7 +151,7 @@ export type FiatExchangeRatesResponse = Infer<
  * For safety, most fields are marked optional and nullable even though it goes against the type in the Price API source code.
  */
 
-export const SpotPriceStruct = object({
+export const SpotPriceStruct = typeStruct({
   id: string(),
   price: min(number(), 0),
   marketCap: optional(nullable(min(number(), 0))),